fix(webapp): fix lazy routing
