+++ apps/lazy-loading-app/src/app/app.module.ts
@@ -4,23 +4,20 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AlertModule } from './alert/alert.module';
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { AppRoutingModule } from './app.routing';
 import { AuthModule } from './auth/auth.module';
 import { ErrorInterceptor } from './helpers/error.interceptor';
 import { JwtInterceptor } from './helpers/jwt.interceptor';
-import { HomeModule } from './home/home.module';
-import { LoginModule } from './login/login.module';
-import { RegisterModule } from './register/register.module';
 
 @NgModule({
   declarations: [AppComponent],
   imports: [
     AlertModule,
     AuthModule,
-    HomeModule,
-    LoginModule,
-    RegisterModule,
+    // HomeModule,
+    // LoginModule,
+    // RegisterModule,
     ReactiveFormsModule,
     BrowserModule,
     HttpClientModule,
