+++ apps/lazy-loading-app/src/app/auth/BUILD.bazel
@@ -4,12 +4,10 @@ package(default_visibility = ["//:__subpackages__"])
 
 ng_module(
     name = "auth",
-    srcs = [
-        "auth.guard.ts",
-        "auth.module.ts",
-        "authentication.service.ts",
-        "index.ts",
-    ],
+    srcs = glob(
+        include = ["*.ts"],
+        exclude = ["*spec.ts"],
+    ),
     tsconfig = "//:tsconfig.json",
     deps = [
         "//libs/api-interface/src:api-interface",
