+++ apps/lazy-loading-app/src/app/home/BUILD.bazel
@@ -17,6 +17,7 @@ ng_module(
         "@npm//@angular/common",
         "@npm//@angular/core",
         "@npm//@angular/forms",
+        "@npm//@angular/router",
         "@npm//rxjs",
         "@npm//tslib",
     ],
