+++ apps/lazy-loading-app/src/app/register/BUILD.bazel
@@ -4,10 +4,10 @@ package(default_visibility = ["//:__subpackages__"])
 
 ng_module(
     name = "register",
-    srcs = [
-        "register.component.ts",
-        "register.module.ts",
-    ],
+    srcs = glob(
+        include = ["*.ts"],
+        exclude = ["*spec.ts"],
+    ),
     assets = glob(["*.html"]),
     tsconfig = "//:tsconfig.json",
     deps = [
